requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
attrs==20.2.0
Automat==20.2.0
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
constantly==15.1.0
cryptography==3.1.1
cssselect==1.1.0
Django==3.1.2
djangorestframework==3.12.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.1
itemloaders==1.0.3
jmespath==0.10.0
lxml==4.6.1
parsel==1.6.0
promise==2.3
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-decouple==3.3
pytz==2020.1
queuelib==1.5.0
Rx==1.6.1
Scrapy==2.4.0
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
Twisted==20.3.0
Unidecode==1.1.1
w3lib==1.22.0
zope.interface==5.1.2
